fix(api): split translation queue worker
