+++ package.json
@@ -101,14 +101,17 @@
     "deploy:cloudflare:console:prod": "wrangler deploy --env=prod",
     "deploy:cloudflare:console:preprod": "wrangler deploy --env=preprod",
     "deploy:cloudflare:console:dev": "wrangler deploy --env=dev",
-    "deploy:cloudflare:prod": "bun run deploy:cloudflare:api:prod && bun run deploy:cloudflare:files:prod && bun run deploy:cloudflare:plugins:prod",
+    "deploy:cloudflare:prod": "bun run deploy:cloudflare:api:prod && bun run deploy:cloudflare:translation:prod && bun run deploy:cloudflare:files:prod && bun run deploy:cloudflare:plugins:prod",
     "deploy:cloudflare:api:prod": "bunx wrangler deploy --config cloudflare_workers/api/wrangler.jsonc --env=prod --minify --outdir dist --upload-source-maps",
+    "deploy:cloudflare:translation:prod": "bunx wrangler deploy --config cloudflare_workers/translation/wrangler.jsonc --env=prod --minify --outdir dist --upload-source-maps",
     "deploy:cloudflare:files:prod": "bunx wrangler deploy --config cloudflare_workers/files/wrangler.jsonc --env=prod --minify",
-    "deploy:cloudflare:preprod": "bun run deploy:cloudflare:api:preprod && bun run deploy:cloudflare:files:preprod && bun run deploy:cloudflare:plugin:preprod",
+    "deploy:cloudflare:preprod": "bun run deploy:cloudflare:api:preprod && bun run deploy:cloudflare:translation:preprod && bun run deploy:cloudflare:files:preprod && bun run deploy:cloudflare:plugin:preprod",
     "deploy:cloudflare:api:preprod": "bunx wrangler deploy --config cloudflare_workers/api/wrangler.jsonc --env=preprod --minify",
+    "deploy:cloudflare:translation:preprod": "bunx wrangler deploy --config cloudflare_workers/translation/wrangler.jsonc --env=preprod --minify",
     "deploy:cloudflare:files:preprod": "bunx wrangler deploy --config cloudflare_workers/files/wrangler.jsonc --env=preprod --minify",
     "deploy:cloudflare:plugin:preprod": "bunx wrangler deploy --config cloudflare_workers/plugin/wrangler.jsonc --env=preprod --minify",
     "deploy:cloudflare:api:dev": "bunx wrangler deploy --config cloudflare_workers/api/wrangler.jsonc --env=alpha --minify",
+    "deploy:cloudflare:translation:dev": "bunx wrangler deploy --config cloudflare_workers/translation/wrangler.jsonc --env=alpha --minify",
     "deploy:cloudflare:files:dev": "bunx wrangler deploy --config cloudflare_workers/files/wrangler.jsonc --env=alpha --minify",
     "deploy:cloudflare:plugin:dev": "bunx wrangler deploy --config cloudflare_workers/plugin/wrangler.jsonc --env=alpha --minify",
     "deploy:cloudflare:plugins:prod": "bun run deploy:cloudflare:plugin_as:prod && bun run deploy:cloudflare:plugin_na:prod && bun run deploy:cloudflare:plugin_eu:prod && bun run deploy:cloudflare:plugin_af:prod && bun run deploy:cloudflare:plugin_oc:prod && bun run deploy:cloudflare:plugin_sa:prod && bun run deploy:cloudflare:plugin_me:prod && bun run deploy:cloudflare:plugin_hk:prod && bun run deploy:cloudflare:plugin_jp:prod",
